Configure chunk splitting.
This way the user doesn't have to download the entire bundle for the
initial page load.
