+++ package.json
@@ -32,6 +32,7 @@
     "react": "^16.6.3",
     "react-apollo": "^2.3.3",
     "react-dom": "^16.6.3",
+    "react-loadable": "^5.5.0",
     "react-router-dom": "^4.3.1",
     "styled-components": "^4.1.3"
   },
@@ -39,6 +40,7 @@
     "@types/graphql": "^14.0.3",
     "@types/react": "^16.7.17",
     "@types/react-dom": "^16.0.11",
+    "@types/react-loadable": "^5.4.2",
     "@types/react-router-dom": "^4.3.1",
     "@types/styled-components": "^4.1.4",
     "awesome-typescript-loader": "^5.2.1",
