+++ public/index.html
@@ -14,6 +14,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script src="/bundle.js" type="text/javascript"></script>
+    <script src="/main.bundle.js" type="text/javascript"></script>
   </body>
 </html>
