--- src/pages/index.ts
@@ -1,5 +1,0 @@
-export { default as Blog } from "./Blog";
-export { default as Contact } from "./Contact";
-export { default as History } from "./History";
-export { default as Home } from "./Home";
-export { default as NotFound } from "./NotFound";
