+++ tsconfig.json
@@ -1,11 +1,15 @@
 {
   "compilerOptions": {
-    "outDir": "./dist/",
+    "outDir": "./dist",
+    "target": "es5",
+    "module": "esnext",
+    "jsx": "react",
+    "declaration": false,
     "sourceMap": true,
     "noImplicitAny": true,
-    "module": "commonjs",
-    "target": "es2015",
-    "jsx": "react",
+    "strict": true,
+    "moduleResolution": "node",
+    "types": ["node", "react", "react-dom"],
     "lib": [
       "es5",
       "es6",
@@ -21,6 +25,5 @@
       "es2015.symbol.wellknown",
       "esnext.asynciterable"
     ]
-  },
-  "include": ["./src/**/*"]
+  }
 }
